+++ src/main/java/com/minecolonies/core/items/ItemSupplyCampDeployer.java
@@ -10,13 +10,13 @@ import com.minecolonies.api.colony.IColony;
 import com.minecolonies.api.colony.IColonyManager;
 import com.minecolonies.api.colony.permissions.Action;
 import com.minecolonies.api.items.ISupplyItem;
-import com.minecolonies.api.items.ModDataComponents;
+import com.minecolonies.api.items.component.ModDataComponents;
+import com.minecolonies.api.items.component.SupplyData;
 import com.minecolonies.api.util.MessageUtils;
 import com.minecolonies.api.util.WorldUtil;
 import com.minecolonies.core.MineColonies;
 import com.minecolonies.core.client.gui.WindowSupplies;
 import com.minecolonies.core.client.gui.WindowSupplyStory;
-import com.minecolonies.core.items.ItemSupplyChestDeployer.SupplyData;
 import net.minecraft.world.entity.player.Player;
 import net.minecraft.world.item.Item;
 import net.minecraft.world.item.ItemStack;
